+++ types/index.d.ts
@@ -28,6 +28,7 @@
 ///<reference path='./merchantInstallment/merchantInstallment.d.ts' />
 ///<reference path='./merchantCard/merchantCard.d.ts' />
 ///<reference path='./split/split.d.ts' />
+///<reference path='./verifiedAccount/verifiedAccount.d.ts' />
 ///<reference path='./splitProfile/splitProfile.d.ts' />
 ///<reference path='./splitReceiver/splitReceiver.d.ts' />